+++ spec/rspec/core/reporter_spec.rb
@@ -3,6 +3,7 @@ module RSpec::Core
     include FormatterSupport
 
     let(:config)   { Configuration.new }
+    let(:world)    { World.new(config) }
     let(:reporter) { Reporter.new config }
     let(:start_time) { Time.now }
     let(:example) { super() }
@@ -283,6 +284,10 @@ module RSpec::Core
 
     describe "#notify_non_example_exception" do
       it "sends a `message` notification that contains the formatted exception details" do
+        if RSpec::Support::Ruby.jruby_9000?
+          pending "RSpec gets `Unable to find matching line from backtrace` on JRuby 9000"
+        end
+
         formatter_out = StringIO.new
         formatter = Formatters::ProgressFormatter.new(formatter_out)
         reporter.register_listener formatter, :message
@@ -300,6 +305,12 @@ module RSpec::Core
           |  divided by 0
           |# #{Metadata.relative_path(__FILE__)}:#{line}
         EOS
+      end
+
+      it "records the fact that a non example failure has occurred" do
+        expect {
+          reporter.notify_non_example_exception(Exception.new, "NonExample Context")
+        }.to change(world, :non_example_failure).from(a_falsey_value).to(true)
       end
     end
   end