+++ spec/rspec/core/suite_hooks_spec.rb
@@ -15,12 +15,23 @@ module RSpec::Core
           }.not_to yield_control
         end
 
-        it 'allows errors in the hook to propagate to the user' do
+        it 'notifies about errors in the hook' do
           RSpec.configuration.__send__(registration_method, :suite) { 1 / 0 }
+
+          expect(RSpec.configuration.reporter).to receive(:notify_non_example_exception).with(
+            ZeroDivisionError, /suite\)` hook/
+          )
+
+          RSpec.configuration.with_suite_hooks { }
+        end
+
+        it 'sets `wants_to_quit` when an error occurs so that the suite does not get run' do
+          RSpec.configuration.__send__(registration_method, :suite) { 1 / 0 }
+          allow(RSpec.configuration.reporter).to receive(:notify_non_example_exception)
 
           expect {
             RSpec.configuration.with_suite_hooks { }
-          }.to raise_error(ZeroDivisionError)
+          }.to change(RSpec.world, :wants_to_quit).from(a_falsey_value).to(true)
         end
 
         it 'runs in the context of an example group' do